update package dependencies and remove unused images, rename main_section to main in components
