+++ frontend/package.json
@@ -23,6 +23,7 @@
     "react-day-picker": "^9.6.3",
     "react-dom": "^18.3.1",
     "sanity": "^3.81.0",
+    "swiper": "^11.2.6",
     "tailwind-merge": "^3.0.1",
     "tailwindcss-animate": "^1.0.7",
     "use-debounce": "^10.0.4"
