+++ frontend/src/components/main/ActivitiesSection.tsx
