+++ frontend/src/components/main/HeroSection.tsx
