+++ frontend/src/components/main/JoinUsSection.tsx
