+++ frontend/src/components/main/SpecializationSection.tsx
